[BM-125] :bug: 상품 관련 Controller -> ApiController로 병합
