+++ src/main/java/com/saiko/bidmarket/product/controller/ProductApiController.java
@@ -3,6 +3,8 @@ package com.saiko.bidmarket.product.controller;
 import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -29,5 +31,11 @@ public class ProductApiController {
       @RequestBody @Valid ProductCreateRequest productCreateRequest) {
     long productId = productService.create(productCreateRequest);
     return ProductCreateResponse.from(productId);
+  }
+
+  @GetMapping("/{id}")
+  @ResponseStatus(HttpStatus.OK)
+  public ProductDetailResponse findById(@PathVariable long id) {
+    return ProductDetailResponse.from(productService.findById(id));
   }
 }
